feat: render IncomingFooter if call is incoming
